refactor: move error to port
