+++ internal/adapter/repository/postgres/order.go
@@ -6,6 +6,7 @@ import (
 
 	sq "github.com/Masterminds/squirrel"
 	"github.com/bagashiz/go-pos/internal/core/domain"
+	"github.com/bagashiz/go-pos/internal/core/port"
 	"github.com/jackc/pgx/v5"
 )
 
@@ -151,7 +152,7 @@ func (or *OrderRepository) GetOrderByID(ctx context.Context, id uint64) (*domain
 		)
 		if err != nil {
 			if err == pgx.ErrNoRows {
-				return domain.ErrDataNotFound
+				return port.ErrDataNotFound
 			}
 			return err
 		}
