+++ internal/adapter/repository/postgres/product.go
@@ -6,6 +6,7 @@ import (
 
 	sq "github.com/Masterminds/squirrel"
 	"github.com/bagashiz/go-pos/internal/core/domain"
+	"github.com/bagashiz/go-pos/internal/core/port"
 	"github.com/jackc/pgx/v5"
 )
 
@@ -81,7 +82,7 @@ func (pr *ProductRepository) GetProductByID(ctx context.Context, id uint64) (*do
 	)
 	if err != nil {
 		if err == pgx.ErrNoRows {
-			return nil, domain.ErrDataNotFound
+			return nil, port.ErrDataNotFound
 		}
 		return nil, err
 	}
