+++ internal/adapter/repository/postgres/user.go
@@ -6,6 +6,7 @@ import (
 
 	sq "github.com/Masterminds/squirrel"
 	"github.com/bagashiz/go-pos/internal/core/domain"
+	"github.com/bagashiz/go-pos/internal/core/port"
 	"github.com/jackc/pgx/v5"
 )
 
@@ -99,7 +100,7 @@ func (ur *UserRepository) GetUserByID(ctx context.Context, id uint64) (*domain.U
 	)
 	if err != nil {
 		if err == pgx.ErrNoRows {
-			return nil, domain.ErrDataNotFound
+			return nil, port.ErrDataNotFound
 		}
 		return nil, err
 	}
