+++ internal/core/port/error.go
@@ -1,4 +1,4 @@
-package domain
+package port
 
 import (
 	"errors"
